@@ -12,7 +12,7 @@ namespace calledudeBot.Chat.Commands
             RequiresMod = true;
         }
 
-        protected override string specialFunc(CommandParameter param)
+        protected override string SpecialFunc(CommandParameter param)
         {
             string response = "You ok there bud? Try again.";
 
