@@ -4,13 +4,13 @@ namespace calledudeBot.Chat.Commands
 {
     public abstract class SpecialCommand<T> : Command where T : CommandParameter
     {
-        protected abstract string specialFunc(T param);
-        public virtual string GetResponse(T param) => specialFunc(param);
+        protected abstract string SpecialFunc(T param);
+        public string GetResponse(T param) => SpecialFunc(param);
     }
 
     public abstract class SpecialCommand : Command
     {
-        protected abstract string specialFunc();
-        public virtual string GetResponse() => specialFunc();
+        protected abstract string SpecialFunc();
+        public string GetResponse() => SpecialFunc();
     }
 }
