@@ -2,10 +2,12 @@
 {
     public sealed class DiscordMessage : Message
     {
-        public ulong Destination { get; set; }
+        public ulong Destination { get; }
 
-        public DiscordMessage(string message) : base(message)
+        public DiscordMessage(string message, string channel, User sender, ulong destination) 
+            : base(message, channel, sender)
         {
+            Destination = destination;
         }
     }
 }
