@@ -20,49 +20,50 @@ namespace calledudeBot
                                             _channelName, _osuNick, _announceChanID, _streamerID;
         private static bool _verifiedBots;
 
-        public static void ProduceBots()
+        public static async Task ProduceBots()
         {
-            getCredentials(); //Makes sure all credentials are at the very least present.
+            GetCredentials(); //Makes sure all credentials are at the very least present.
 
             Bot.TestRun = true;
-            while (!VerifyCredentials())
+            while (!await VerifyCredentials())
             {
-                getMissingCredentials();
-                tryLoadCredentials();
+                GetMissingCredentials();
+                TryLoadCredentials();
             }
             Bot.TestRun = false;
         }
 
-        public static List<Bot> GetVerifiedBots(out DiscordBot discordBot, out TwitchBot twitchBot, out OsuBot osuBot)
+        public static IReadOnlyCollection<Bot> GetVerifiedBots(out DiscordBot discordBot, out TwitchBot twitchBot, out OsuBot osuBot)
         {
-            if (!_verifiedBots) throw new InvalidOperationException("You're not allowed to call this method before all bots have been verified.");
+            if (!_verifiedBots)
+                throw new InvalidOperationException("You're not allowed to call this method before all bots have been verified.");
+
             discordBot = new DiscordBot(_discordToken, ulong.Parse(_announceChanID), ulong.Parse(_streamerID));
             osuBot = new OsuBot(_osuIRCtoken, _osuNick);
-            twitchBot = new TwitchBot(_twitchIRCtoken, _osuAPIToken, _osuNick, _botNick, _channelName);
-            return new List<Bot> { discordBot, twitchBot, osuBot };
+            twitchBot = new TwitchBot(_twitchIRCtoken, _osuAPIToken, _osuNick, _botNick, _channelName, osuBot);
+
+            return new Bot[] { discordBot, twitchBot, osuBot };
         }
 
         //Returns a boolean after running every single bot through the verify function
-        private static bool VerifyCredentials()
+        private static async Task<bool> VerifyCredentials()
         {
-            bool discToken = false, twitchToken = false, osuToken = false;
-            Parallel.Invoke(
-                () => discToken = VerifyBot<DiscordBot>(),
-                () => twitchToken = VerifyBot<TwitchBot>(),
-                () => osuToken = VerifyBot<OsuBot>());
+            var discToken = Task.Run(() => VerifyBot<DiscordBot>());
+            var twitchToken = Task.Run(() => VerifyBot<TwitchBot>());
+            var osuToken = Task.Run(() => VerifyBot<OsuBot>());
 
-            return _verifiedBots = discToken && twitchToken && osuToken;
+            var results = await Task.WhenAll(discToken, twitchToken, osuToken);
+            return _verifiedBots = results.All(x => x);
         }
 
-        private static bool VerifyBot<T>() where T : Bot
+        public static async Task<bool> VerifyBot<T>() where T : Bot
         {
-            using (Bot bot = getBotInstance<T>())
+            using (Bot bot = GetBotInstance<T>())
             {
-                if (bot == null) return false;
                 bool success = false;
                 try
                 {
-                    bot.Start().GetAwaiter().GetResult();
+                    await bot.Start();
                     success = true; //Will only be set if bot.Start() does not throw an exception
                 }
                 catch (HttpException)
@@ -76,12 +77,22 @@ namespace calledudeBot
                     {
                         _creds.Remove("osuIRC");
                         _creds.Remove("OsuNick");
-                        bot.TryLog("Invalid token and/or osu!-nickname");
+                        bot.TryLog("Invalid token and/or osu!-nickname.");
                     }
                     else
                     {
                         _creds.Remove("TwitchIRC");
-                        bot.TryLog("Invalid token");
+                        bot.TryLog("Invalid token.");
+                    }
+                }
+                catch (TimeoutException)
+                {
+                    //For whatever reason, Twitch IRC doesn't verify the nickname
+                    //Nor does it verify that the channel actually exists
+                    if (bot is TwitchBot)
+                    {
+                        _creds.Remove("ChannelName");
+                        _creds.Remove("BotNick");
                     }
                 }
                 catch (WebException)
@@ -89,7 +100,7 @@ namespace calledudeBot
                     bot.TryLog("Invalid osu! API token.");
                     _creds.Remove("osuAPI");
                 }
-                catch (ArgumentException)
+                catch (ArgumentNullException)
                 {
                     _creds.Remove("OsuNick");
                     bot.TryLog("Invalid osu!-nickname.");
@@ -97,13 +108,13 @@ namespace calledudeBot
                 finally
                 {
                     bot.TryLog($"Verification {(success ? "SUCCESS." : "FAIL.")}");
-                    bot.Logout().GetAwaiter().GetResult();
+                    await bot.Logout();
                 }
                 return success;
             }
         }
 
-        private static Bot getBotInstance<T>() where T : Bot
+        private static Bot GetBotInstance<T>() where T : Bot
         {
             if (typeof(T) == typeof(DiscordBot))
             {
@@ -124,7 +135,7 @@ namespace calledudeBot
             }
             else if (typeof(T) == typeof(TwitchBot))
             {
-                return new TwitchBot(_twitchIRCtoken, _osuAPIToken, _osuNick, _botNick, _channelName);
+                return new TwitchBot(_twitchIRCtoken, _osuAPIToken, _osuNick, _botNick.ToLower(), _channelName.ToLower(), null);
             }
             else
             {
@@ -132,7 +143,7 @@ namespace calledudeBot
             }
         }
 
-        private static bool tryLoadCredentials()
+        public static bool TryLoadCredentials()
         {
             return _creds.TryGetValue("BotNick", out _botNick)
                 && _creds.TryGetValue("ChannelName", out _channelName)
@@ -145,7 +156,7 @@ namespace calledudeBot
                 && _creds.TryGetValue("StreamerID", out _streamerID);
         }
 
-        private static void getMissingCredentials()
+        private static void GetMissingCredentials()
         {
             Dictionary<string, string> required = new Dictionary<string, string>
             {
@@ -167,14 +178,14 @@ namespace calledudeBot
             Type t = typeof(CredentialChecker);
             foreach (string s in missing)
             {
-                MethodInfo m = t.GetMethod("get" + s, BindingFlags.NonPublic | BindingFlags.Static);
+                MethodInfo m = t.GetMethod("Get" + s, BindingFlags.NonPublic | BindingFlags.Static);
                 m.Invoke(null, null);
             }
 
             File.Create(_credFile).Close();
-            foreach (KeyValuePair<string, string> k in _creds)
+            foreach (var kvp in _creds)
             {
-                File.AppendAllText(_credFile, k.Key + " " + k.Value + Environment.NewLine);
+                File.AppendAllText(_credFile, $"{kvp.Key} {kvp.Value} {Environment.NewLine}");
             }
 
             Console.WriteLine("Alright! We're all done. Let's go! :)");
@@ -182,29 +193,30 @@ namespace calledudeBot
             Console.Clear();
         }
 
-        private static void getCredentials()
+        public static void GetCredentials()
         {
             if (File.Exists(_credFile))
             {
                 _creds = File.ReadAllLines(_credFile)
                     .Distinct()
-                    .Where(p => p.Trim().Split(' ').Length == 2)
-                    .ToDictionary(key => key.Split(' ')[0], val => val.Split(' ')[1]);
+                    .Select(x => x.Trim().Split(' '))
+                    .Where(p => p.Length == 2)
+                    .ToDictionary(key => key[0], val => val[1]);
             }
             else
             {
                 File.Create(_credFile).Close();
                 _creds = new Dictionary<string, string>();
             }
-            while (!tryLoadCredentials())
+            while (!TryLoadCredentials())
             {
-                getMissingCredentials();
-                getCredentials(); //Lets try the newly entered credentials again.
+                GetMissingCredentials();
+                GetCredentials(); //Lets try the newly entered credentials again.
             }
         }
 
         #region Setup
-        private static void openWebsite(int delay, string url) //Opens a website in a non-blocking manner after the specified delay
+        private static void OpenWebsite(int delay, string url) //Opens a website in a non-blocking manner after the specified delay
         {
             Task.Run(() =>
             {
@@ -214,7 +226,7 @@ namespace calledudeBot
             });
         }
 
-        private static bool getConfirmation()
+        private static bool GetConfirmation()
         {
             ConsoleKey c;
             do c = Console.ReadKey(true).Key;
@@ -224,17 +236,17 @@ namespace calledudeBot
             return c == ConsoleKey.Y;
         }
 
-        private static void getBotNick()
+        private static void GetBotNick()
         {
             Console.Write("What will the username be of your bot be?: ");
             _botNick = Console.ReadLine();
             _creds.Add("BotNick", _botNick);
         }
 
-        private static void getChannelName()
+        private static void GetChannelName()
         {
             Console.Write("Will you be using your personal twitch account as the 'bot'? Y/N: ");
-            if (getConfirmation())
+            if (GetConfirmation())
             {
                 _channelName = "#" + _botNick;
             }
@@ -246,64 +258,64 @@ namespace calledudeBot
             _creds.Add("ChannelName", _channelName);
         }
 
-        private static void getOsuNick()
+        private static void GetOsuNick()
         {
             Console.Write("What's your osu! username?: ");
             _osuNick = Console.ReadLine();
             _creds.Add("OsuNick", _osuNick);
         }
 
-        private static void getAnnounceChannelID()
+        private static void GetAnnounceChannelID()
         {
             Console.Write("What channel on your discord server do you want the announcements to be made on? (Long number): ");
-            openWebsite(4000, "https://support.discordapp.com/hc/en-us/articles/206346498-Where-can-I-find-my-User-Server-Message-ID");
+            OpenWebsite(4000, "https://support.discordapp.com/hc/en-us/articles/206346498-Where-can-I-find-my-User-Server-Message-ID");
 
             _announceChanID = Console.ReadLine();
             _creds.Add("AnnounceChannelID", _announceChanID);
         }
 
-        private static void getDiscordToken()
+        private static void GetDiscordToken()
         {
             Console.WriteLine("First you need to create a bot/app on the discord developer website.");
             Console.Write("Discord token: ");
-            openWebsite(3000, "https://discordapp.com/developers/applications/me");
+            OpenWebsite(3000, "https://discordapp.com/developers/applications/me");
 
             _discordToken = Console.ReadLine();
             Console.Write("Have you figured out how to get the bot to join your own channel? Y/N: ");
 
-            if (!getConfirmation())
+            if (!GetConfirmation())
             {
                 Console.Write("Alright, enter the bots Client-ID here: ");
-                openWebsite(0, $"https://discordapp.com/oauth2/authorize?client_id={Console.ReadLine()}&scope=bot");
+                OpenWebsite(0, $"https://discordapp.com/oauth2/authorize?client_id={Console.ReadLine()}&scope=bot");
             }
             _creds.Add("DiscordToken", _discordToken);
         }
 
-        private static void getTwitchIRC()
+        private static void GetTwitchIRC()
         {
             Console.Write("Twitch IRC token: ");
-            openWebsite(2000, "http://www.twitchapps.com/tmi/");
+            OpenWebsite(2000, "http://www.twitchapps.com/tmi/");
             _twitchIRCtoken = Console.ReadLine();
             _creds.Add("TwitchIRC", _twitchIRCtoken);
         }
 
-        private static void getosuIRC()
+        private static void GetosuIRC()
         {
             Console.Write("osu! IRC password: ");
-            openWebsite(2000, "https://osu.ppy.sh/p/irc");
+            OpenWebsite(2000, "https://osu.ppy.sh/p/irc");
             _osuIRCtoken = Console.ReadLine();
             _creds.Add("osuIRC", _osuIRCtoken);
         }
 
-        private static void getosuAPI()
+        private static void GetosuAPI()
         {
             Console.Write("osu! API key: ");
-            openWebsite(2000, "https://osu.ppy.sh/p/api");
+            OpenWebsite(2000, "https://osu.ppy.sh/p/api");
             _osuAPIToken = Console.ReadLine();
             _creds.Add("osuAPI", _osuAPIToken);
         }
 
-        private static void getStreamerID()
+        private static void GetStreamerID()
         {
             Console.Write("YOUR User-ID (On Discord): ");
             _streamerID = Console.ReadLine();
