@@ -10,9 +10,9 @@ using System.Timers;
 
 namespace calledudeBot.Services
 {
-    public class StreamMonitor : IDisposable
+    public sealed class StreamMonitor : IDisposable
     {
-        private readonly IGuildUser _streamer;
+        private IGuildUser _streamer;
         private readonly OBSWebsocket _obs;
         private readonly Timer _streamStatusTimer;
         private readonly DiscordSocketClient _client;
@@ -21,7 +21,7 @@ namespace calledudeBot.Services
         public bool IsStreaming { get; private set; }
         public DateTime StreamStarted { get; private set; }
 
-        public StreamMonitor(ulong streamerID, ulong announceChanID, DiscordSocketClient client)
+        private StreamMonitor(ulong announceChanID, DiscordSocketClient client)
         {
             _client = client;
             _obs = new OBSWebsocket();
@@ -32,8 +32,32 @@ namespace calledudeBot.Services
             _streamStatusTimer.Elapsed += CheckDiscordStatus;
 
             _announceChannel = _client.GetChannel(announceChanID) as ITextChannel;
-            _streamer = _announceChannel.Guild.GetUserAsync(streamerID)
-                                            .GetAwaiter().GetResult();
+        }
+
+        public static async Task<StreamMonitor> Create(ulong announceChanID, ulong streamerID, DiscordSocketClient client)
+        {
+            var monitor = new StreamMonitor(announceChanID, client);
+            return await monitor.TryInitialize(streamerID) 
+                ? monitor
+                : throw new ArgumentException("One or more of the arguments were invalid.");
+        }
+
+        private async Task<bool> TryInitialize(ulong streamerID)
+        {
+            if (_announceChannel == null)
+            {
+                Log("Invalid channel. Will not announce when stream goes live.");
+                return false;
+            }
+
+            _streamer = await _announceChannel.Guild.GetUserAsync(streamerID);
+            if (_streamer == null)
+            {
+                Log("Invalid StreamerID. Could not find user.");
+                return false;
+            }
+
+            return true;
         }
 
         private void Log(string message)
@@ -54,7 +78,9 @@ namespace calledudeBot.Services
             }
 
             //Trying 5 times just in case.
-            if (Enumerable.Range(1, 5).Select(_ => _obs.Connect("ws://localhost:4444")).All(x => !x))
+            if (Enumerable.Range(1, 5)
+                .Select(_ => _obs.Connect("ws://localhost:4444"))
+                .All(x => !x))
             {
                 Log("You need to install the obs-websocket plugin for OBS and configure it to run on port 4444.");
                 await Task.Delay(3000);
@@ -91,9 +117,10 @@ namespace calledudeBot.Services
                 //In that case we assume that the bot has been restarted (for whatever reason)
                 if (!messages.Any(m =>
                     m.Author.Id == _client.CurrentUser.Id
-                    && m.Content == msg
+                    && m.Content.Equals(msg)
                     && StreamStarted - m.Timestamp < TimeSpan.FromMinutes(3)))
                 {
+                    Log($"Streamer went live. Sending announcement to #{_announceChannel.Name}");
                     await _announceChannel.SendMessageAsync(msg);
                 }
             }
